fix: home, mypage css 수정
