+++ src/components/MyPageStudyTime.jsx
@@ -1,14 +1,15 @@
 import React from 'react';
-import { useState } from 'react';
-import { useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import changeTimeSimpleForm from './changeTimeSimpleForm';
 import changeTimeForm from './changeTimeForm';
 import styles from '../css/mypage.module.css';
 import axios from 'axios';
 import { get_studytime } from '../app/slice/timeTimerSlice';
+import styled from 'styled-components';
 
 const MyPageStudyTime = () => {
+    const color = useSelector((state) => state.color);
     const date = new Date().getTime();
     const dispatch = useDispatch();
     const [totalStudyTime, setTotalStudyTime] = useState(0);
@@ -58,13 +59,30 @@ const MyPageStudyTime = () => {
 
     return (
         <div className={styles.container}>
-            <div className={styles.aboveArea}>
+            <Gradient color={color} className={styles.aboveArea}>
                 <div className={styles.title}>오늘 공부시간</div>
                 {changeTimeForm(second, styles.todayStudyTime)}
                 <div className={styles.text}>누적 공부시간은 {changeTimeSimpleForm(totalStudyTime, styles.totalTime)}이에요</div>
-            </div>
+            </Gradient>
         </div>
     );
 };
 
 export default MyPageStudyTime;
+
+const Gradient = styled.div`
+    width: 100%;
+    height: 14rem;
+    background-image: ${(props) => {
+        switch (props.color) {
+            case 'green':
+                return 'linear-gradient(to bottom, var(--neutral-20), #3b4f4b 34%, #558d71 74%, var(--primary-60))';
+            case 'blue':
+                return 'linear-gradient(to bottom, var(--neutral-20), #445364 48%, #5f809b 75%, var(--tertiary-60))';
+            case 'red':
+                return 'linear-gradient(to bottom, var(--neutral-20), #4f3f40 43%, #996153 73%, var(--secondary-60))';
+            default:
+                return 'none';
+        }
+    }};
+`;
