+++ src/components/common/Layout.jsx
@@ -4,9 +4,7 @@ import styled from 'styled-components';
 import '../../css/color.css';
 
 const Layout = (props) => {
-    const color = useSelector((state) => state.color);
-    console.log(color);
-    return <StyledLayout color={color}>{props.children}</StyledLayout>;
+    return <StyledLayout>{props.children}</StyledLayout>;
 };
 
 const StyledLayout = styled.div`
@@ -15,18 +13,6 @@ const StyledLayout = styled.div`
     justify-content: center;
     align-items: center;
     background-color: var(--neutral-20);
-    background-image: ${(props) => {
-        switch (props.color) {
-            case 'green':
-                return 'linear-gradient(to bottom, var(--neutral-20), #3b4f4b 34%, #558d71 74%, var(--primary-60))';
-            case 'blue':
-                return 'linear-gradient(to bottom, var(--neutral-20), #445364 48%, #5f809b 75%, var(--tertiary-60))';
-            case 'red':
-                return 'linear-gradient(to bottom, var(--neutral-20), #4f3f40 43%, #996153 73%, var(--secondary-60))';
-            default:
-                return 'none';
-        }
-    }};
     @media (max-width: 1200px) {
         width: 100%;
     }
